Code cleanup.

+++ BinReader/Pool.cs
@@ -1,5 +1,3 @@
-using System.Reflection;
-
 namespace BinReader
 {
     //a representation of the entire pool containing all of its soldiers
@@ -14,17 +12,16 @@ namespace BinReader
             soldiers.Add(soldier);
         }
 
-        //print out the entire contents of the pool (for now it's a seat warmer, need to fill out)
-        //will also need to be renovated/replaced by a PrintPoolToFile command
-        public void PrintPool()
+
+        //prints the pool to the console, mainly used for debugging the actual printing process
+        //except for file handling, is otherwise identical to PrintPoolToFile
+        public void PrintPoolToConsole()
         {
-            //Console.WriteLine("POOL [FILE NAME]\n\n");    //straight up don't feel like doing this rn
             Console.WriteLine("FIRST NAME | LAST NAME |  NICKNAME   |   CLASS   |GENDER|  NATIONALITY  ");
             Console.WriteLine("-----------|-----------|-------------|-----------|------|---------------");
-            //now print all the soldier details
+            
             foreach (Soldier soldier in soldiers)
             {
-                //awful
                 Console.WriteLine(soldier.FirstName.PadRight(11) + "|"
                     + soldier.LastName.PadRight(11) + "|"
                     + soldier.NickName.PadRight(13) + "|"
@@ -41,10 +38,7 @@ namespace BinReader
             string filePath = file.Remove(file.Length - 3, 3) + "txt";
 
             //if the file does exist, kill it (sorry file)
-            if (File.Exists(filePath))
-            {
-                File.Delete(filePath);
-            }
+            if (File.Exists(filePath)) { File.Delete(filePath); }
 
             //create file
             using (FileStream fs = File.Create(filePath))
@@ -52,13 +46,11 @@ namespace BinReader
                 //create StreamWriter for it
                 using (StreamWriter sw = new StreamWriter(fs))
                 {
-                    //copy the writing code and pray
                     sw.WriteLine("FIRST NAME | LAST NAME |  NICKNAME   |   CLASS   |GENDER|  NATIONALITY  ");
                     sw.WriteLine("-----------|-----------|-------------|-----------|------|---------------");
-                    //now print all the soldier details
+                    
                     foreach (Soldier soldier in soldiers)
                     {
-                        //awful
                         sw.WriteLine(soldier.FirstName.PadRight(11) + "|"
                             + soldier.LastName.PadRight(11) + "|"
                             + soldier.NickName.PadRight(13) + "|"
